Criteria. Move Expression related classes into separate (expression) package
