+++ criteria/common/src/org/immutables/criteria/constraints/CollectionCriteria.java
@@ -1,6 +1,9 @@
 package org.immutables.criteria.constraints;
 
 import org.immutables.criteria.DocumentCriteria;
+import org.immutables.criteria.expression.Expression;
+import org.immutables.criteria.expression.Expressions;
+import org.immutables.criteria.expression.Operators;
 
 import java.util.Objects;
 import java.util.function.UnaryOperator;
