+++ criteria/common/src/org/immutables/criteria/expression/Expressional.java
@@ -1,4 +1,4 @@
-package org.immutables.criteria.constraints;
+package org.immutables.criteria.expression;
 
 
 /**
