+++ criteria/common/src/org/immutables/criteria/expression/Operator.java
@@ -1,4 +1,4 @@
-package org.immutables.criteria.constraints;
+package org.immutables.criteria.expression;
 
 /**
  * Marker interface for now.
