+++ criteria/common/src/org/immutables/criteria/expression/Operators.java
@@ -1,4 +1,4 @@
-package org.immutables.criteria.constraints;
+package org.immutables.criteria.expression;
 
 public enum Operators implements Operator {
