+++ criteria/common/test/org/immutables/criteria/ExpressionAsStringTest.java
@@ -1,7 +1,7 @@
 package org.immutables.criteria;
 
-import org.immutables.criteria.constraints.DebugExpressionVisitor;
-import org.immutables.criteria.constraints.Expressional;
+import org.immutables.criteria.expression.DebugExpressionVisitor;
+import org.immutables.criteria.expression.Expressional;
 import org.junit.Assert;
 import org.junit.Test;
