+++ criteria/inmemory/src/org.immutables.criteria.inmemory/InMemoryExpressionEvaluator.java
@@ -1,14 +1,14 @@
 package org.immutables.criteria.inmemory;
 
 import com.google.common.base.Preconditions;
-import org.immutables.criteria.constraints.Call;
-import org.immutables.criteria.constraints.Constant;
-import org.immutables.criteria.constraints.Expression;
-import org.immutables.criteria.constraints.ExpressionVisitor;
-import org.immutables.criteria.constraints.Expressions;
-import org.immutables.criteria.constraints.Operator;
-import org.immutables.criteria.constraints.Operators;
-import org.immutables.criteria.constraints.Path;
+import org.immutables.criteria.expression.Call;
+import org.immutables.criteria.expression.Constant;
+import org.immutables.criteria.expression.Expression;
+import org.immutables.criteria.expression.ExpressionVisitor;
+import org.immutables.criteria.expression.Expressions;
+import org.immutables.criteria.expression.Operator;
+import org.immutables.criteria.expression.Operators;
+import org.immutables.criteria.expression.Path;
 
 import javax.annotation.Nullable;
 import java.lang.reflect.Field;
