+++ criteria/mongo/src/org/immutables/criteria/mongo/MongoVisitor.java
@@ -11,13 +11,13 @@ import org.bson.BsonWriter;
 import org.bson.codecs.Encoder;
 import org.bson.codecs.EncoderContext;
 import org.bson.codecs.configuration.CodecRegistry;
-import org.immutables.criteria.constraints.Call;
-import org.immutables.criteria.constraints.Constant;
-import org.immutables.criteria.constraints.Expression;
-import org.immutables.criteria.constraints.ExpressionVisitor;
-import org.immutables.criteria.constraints.Operator;
-import org.immutables.criteria.constraints.Operators;
-import org.immutables.criteria.constraints.Path;
+import org.immutables.criteria.expression.Call;
+import org.immutables.criteria.expression.Constant;
+import org.immutables.criteria.expression.Expression;
+import org.immutables.criteria.expression.ExpressionVisitor;
+import org.immutables.criteria.expression.Operator;
+import org.immutables.criteria.expression.Operators;
+import org.immutables.criteria.expression.Path;
 
 import java.util.List;
 import java.util.Objects;
