+++ value-processor/src/org/immutables/value/processor/Criteria.generator
@@ -44,15 +44,15 @@ import org.immutables.criteria.constraints.ObjectCriteria;
 import org.immutables.criteria.constraints.StringCriteria;
 import org.immutables.criteria.constraints.BooleanCriteria;
 import org.immutables.criteria.constraints.ComparableCriteria;
-import org.immutables.criteria.constraints.Expression;
-import org.immutables.criteria.constraints.Expressional;
-import org.immutables.criteria.constraints.ExpressionVisitor;
-import org.immutables.criteria.constraints.Expressions;
-import org.immutables.criteria.constraints.Constant;
-import org.immutables.criteria.constraints.Operators;
-import org.immutables.criteria.constraints.Operator;
-import org.immutables.criteria.constraints.Path;
-import org.immutables.criteria.constraints.Call;
+import org.immutables.criteria.expression.Expression;
+import org.immutables.criteria.expression.Expressional;
+import org.immutables.criteria.expression.ExpressionVisitor;
+import org.immutables.criteria.expression.Expressions;
+import org.immutables.criteria.expression.Constant;
+import org.immutables.criteria.expression.Operators;
+import org.immutables.criteria.expression.Operator;
+import org.immutables.criteria.expression.Path;
+import org.immutables.criteria.expression.Call;
 
 import java.util.Objects;
